Update LICENSE
The link to the license on the pgEdge website is incorrect. Personally I like the URL you pointed at more than the actual url, so maybe it'd be better to fix the url in the website (with a 301 from the old to the new url), although if you are changing the website url, I'd probably also suggesting putting the 1.0 version number in the url, but lieu of any of that, the docs should point to the right place.