Update snowflake-docker.yml
I accidentally merged the updates to our GH Actions into the main branch when I committed the README file updates, so am updating the workflow to execute on push to the main branch; I'll still make a couple of updates to the workflow, but will resume work in another branch to isolate the changes.  At this point, the changes are sound.