Introduce regression tests (spoc-220).
It appears that there has been a little mess with the snowflake tests.
This commit aims to fix some 'low-hanging fruits' in the Makefile. Also, here
we add the 'conversion test that is just an intermediate step on a way to let
the convert_sequence_to_snowflake function process correctly different types
of sequences, including IDENTITY constraints.