Error on conversion of IDENTITY column
Second step of improvement on the snowflake conversion function.
Detect the IDENTITY flag and raise an error to avoid conversion that, for now,
just breaks the sequence.

TODO:
Snowflake values are not stable for use in regression tests. Do we need
to develop a tool to verify the accuracy of these values?