Process sequences used as an IDENTITY (spoc-220).
Transform IDENTITY column definition to DEFAULT Snowflake's nextval.
This commit is the third step on a way to change the conversion routine to
process IDENTITY columns correctly. As a side effect, it eases the IDENTITY
ALWAYS restriction to DEFAULT.

NOTE:
The Postgres core may be modified slightly to allow users to utilise
an alternative source of identity columns instead of relying solely
on nextval. Or should we implement our own Snowflake data type and handle
everything there?